requirements.txt
afmparser==1.0.3
alembic==1.18.3
aniso8601==10.0.1
antlr4-python3-runtime==4.13.1
astutils==0.0.6
async-generator==1.10
attrs==25.4.0
Authlib==1.6.7
autoflake==2.3.1
beautifulsoup4==4.14.3
bidict==0.23.1
black==26.1.0
bleach==6.3.0
blinker==1.9.0
brotli==1.2.0
bs4==0.0.2
cachelib==0.13.0
captcha==0.7.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
coverage==7.13.4
croniter==6.0.0
cryptography==46.0.4
dd==0.5.7
dnspython==2.8.0
docker==7.1.0
elastic-transport==8.17.1
elasticsearch==8.18.1
email-validator==2.3.0
Faker==40.4.0
fakeredis==2.33.0
flake8==7.3.0
Flake8-pyproject==1.2.4
flamapy==2.1.0.dev1
flamapy-bdd==2.1.0.dev1
flamapy-fm==2.1.0.dev1
flamapy-fw==2.1.0.dev1
flamapy-sat==2.1.0.dev1
flasgger==0.9.7.2.dev2
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fmfactlabel==1.8.1
gevent==25.9.1
geventhttpclient==2.3.7
graphviz==0.21
greenlet==3.3.1
gunicorn==25.0.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaitaistruct==0.11
locust==2.43.3.dev2
locust-cloud==1.30.0
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mistune==3.2.0
msgpack==1.1.2
msgspec==0.20.0
mypy_extensions==1.1.0
networkx==3.6.1
outcome==1.3.0.post0
packaging==26.0
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
ply==3.10
psutil==7.2.2
pyasn1==0.6.2
pycodestyle==2.14.0
pycparser==3.0
pydot==4.0.1
pyflakes==3.4.0
Pygments==2.19.2
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.3.2
PySocks==1.7.1
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-sat==0.1.8.dev17
python-socketio==5.16.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rq==2.6.1
selenium==4.40.0
selenium-wire==5.1.0
setuptools==81.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.1.0b1
SQLAlchemy-Utils==0.42.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
uvlparser==2.0.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.5
wheel==0.46.3
wsproto==1.3.2
WTForms==3.2.1
detect-secrets==1.5.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
zstandard==0.25.0
